remove warning

+++ JLOSChina/Entities/OSCActivityEntity.h
@@ -9,7 +9,7 @@
 #import "JLNimbusEntity.h"
 
 @interface OSCActivityEntity : JLNimbusEntity
-@property (nonatomic, assign) unsigned long authorId;
+@property (nonatomic, assign) long long authorId;
 @property (nonatomic, copy) NSString* authorName;
 @property (nonatomic, copy) NSString* avatarUrl;
 @end
